// src/store/types.ts
export const GET_WEATHER = "GET_WEATHER";
export const SET_LOADING = "SET_LOADING";
export const SET_ERROR = "SET_ERROR";
export const SET_ALERT = "SET_ALERT";
export const SET_ERROR_FALSE = "SET_ERROR_FALSE"

export interface Weather {
  description: string;
  icon: string;
  id: number;
  main: string;
}

export interface WeatherData {
  coord: {
    lon: number;
    lat: number;
  };
  weather: Weather[];
  base: string;
  main: {
    temp: number;
    feels_like: number;
    temp_min: number;
    temp_max: number;
    pressure: number;
    humidity: number;
  };
  visibility: number;
  wind: {
    speed: number;
    deg: number;
  };
  clouds: {
    all: number;
  };
  dt: number;
  sys: {
    type: number;
    id: number;
    message: number;
    country: string;
    sunrise: number;
    sunset: number;
  };
  timezone: number;
  id: number;
  name: string;
  cod: number;
}


export interface WeatherError{
    cod: string;
    message: string;
}

export interface Error{
  isError: boolean
  message: string
}

export interface WeatherState{
    data: WeatherData | null;
    loading: boolean;
    error: {
      isError: boolean
      message?: string
    } | null;
}

interface GetWeatherAction {
    type: typeof GET_WEATHER;
    payload: WeatherData;
}

interface SetLoadingAction {
    type: typeof SET_LOADING;
}

interface SetErrorAction {
    type: typeof SET_ERROR;
    payload: string
}

interface SetErrorFalseAction {
  type: typeof SET_ERROR_FALSE
}

export type WeatherAction = GetWeatherAction | SetLoadingAction | SetErrorAction | SetErrorFalseAction

export interface AlertAction{
    type: typeof SET_ALERT;
    payload: string;
}

export interface AlertState{
    message: string;
}